requirements.txt
streamlit==1.48.0
streamlit-option-menu==0.4.0
pillow==11.3.0
torch==2.8.0
torchvision==0.23.0
pandas==2.3.1
numpy==2.2.6
ultralytics==8.3.176
huggingface-hub==0.34.4
opencv-contrib-python-headless==4.10.0.84
opencv-python-headless==4.12.0.88
openpyxl==3.1.3